Enhanced Business Logic

+++ src/main/java/com/z20let/mitigia/model/Project.java
@@ -20,7 +20,7 @@ public class Project {
     private LocalDate endDate;
     private Integer startMileage;
     private Integer endMileage;
-    private Long carbonEmission;
+    private Double carbonEmission;
     @ManyToOne(cascade = {CascadeType.PERSIST, CascadeType.MERGE})
     @JoinColumn(name = "vehicle_id")
     Vehicle vehicle;
